[SYCL] Improve testing of host-task (#2540)
This patch introduces two use-cases for host-task implementation.

The issue of the first use-case (`test3`) was about performance and a long-run/wait of host-taske depening on another host-task. This issue rooted in lots of enqueue calls for blocked command and was fixed in bc8f0a4 (#2347).

The second use-case shows a deadlock situation when host-task `A` depends on the blocked one `B` via call `handler::depends_on()`. The `A` host-task won't be enqueued as it's `EmptyCommand` is in blocked state.
The test for this use-case is disable now and will be enabled when the deadlock is fixed.

The deadlock will be fixed in follow-up patch.

This is cherry-pick with tests for CMPLRLLVM-23210.

Change-Id: I77cd6165c1972d6a19312919848fb65ad06726c6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/274805
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>
Tested-by: Sergey Kanaev <sergey.kanaev@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>