[Paropt][Fortran][Target] Add support for dope vectors (F90_DV) in use_device_ptr clause on 'target variant dispatch' constructs."
The change handles cases where a variant function expects a dope
vector to be passed through the parameter list, instead of a raw
pointer to the underlying data:

  subroutine foo_gpu(a)
    integer :: a(:)
    !$omp target is_device_ptr(a)
    ...
  end subroutine

  ...
  !$omp target variant dispatch use_device_ptr(b)
  call foo(b)
  ...

"use_device_ptr" on the dope vector "b" is handled by:
* creating a host copy of the dope vector struct, with its addr0 field
  containing a device pointer to the data. This is done when the IR
  contains an "F90_DV" modifier on the use_device_ptr clause.

"is_device_ptr" on the dope vector "a" is handled by:
* treating it as firstprivate. So the full dope vector struct, which
  already contains the device address of the data in the addr0 field
  (because of use_device_ptr(b)), is copied as is to the target region,
  and references to the addr0 field automatically access device memory.
  (Existing code, not part of this commit).

This commit only adds support for F90_DV modifier on use_device_ptr for
target variant dispatch constructs. The above behavior was already present
for target data.

Note that this new support is not compatible with the deprecated flag
"-vpo-paropt-use-raw-dev-ptr=false".

Change-Id: I4c9a2e34c255cf1ce6a018c7c4ca296c01fa1828
JIRA: CMPLRLLVM-23353
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/273588
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>