CMPLRLLVM-22390: Fix for compfail in 627.cam4 and SCCP
This change set fixes a compfail caused by the SCCP transformation.
The issue is that an array access can be converted as a pointer inside
a call to a function. This casted pointer can then be passed to a
callback. If the constant propagation finds a GEP that could be
optimized then it will try to generate a GEP with incorrect types
producing a compfail.

NOTE: This change set fix the build issue, not the run failure.
The runfail is addressed in CMPLRLLVM-18427.

Change-Id: Ic3d5a2e031a31b1455579a3ef3a9524efe9c0103
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/274481
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>