Correct InstCombinePass constructor call in the new pass manager pipeline
ExpensiveCombines option has been removed from the InstCombinePass
constructor a while ago (please see https://reviews.llvm.org/D76540,
https://git-amr-2.devtools.intel.com/gerrit/#/c/240139/ for details),
but llvm/lib/Passes/PassBuilder.cpp has not been adjusted for this
change. This patch fixes this problem, but I am not sure how to test
this change, so it does not include a LIT test for this change.

Change-Id: Ia54ef4db392aa057d6a43ba0fe37645307ef59c0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/274609
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>