[SYCL] Release commands with no dependencies after they're enqueued (#2492)
Presently, commands that do not have memory dependencies are only released if .wait() is called. If it is not, several resources (event, queue, command, kernel) are held onto.
This fix deletes stores that command event in the USMEvents ( so it is 'owned') and then deletes the command itself. Added a lit test to verify resource release.

Signed-off-by: Chris Perkins <chris.perkins@intel.com>
Change-Id: I02ae71fdeab336744aa9fdd6604ecba2341d99d5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/276781
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>