[SYCL] Relax test to work in Win32 environment.
Our downstream runs this test on a 32 bit windwos machine, so relax the
test slightly to allow it to pass there.

The problem is that the type of 'size_t' (the array index type) is
different (it is unsigned int instead of unsigned long/unsigned long
long), so this changes the test to allow all 3.

Change-Id: Icb453ce512e80e16ab91aaa44d15edad0d7b467d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/274862
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>