CMPLRLLVM-23243: Include the default case when multiversioning a virtual call and one of the target or the caller is a libfunc or wasn't internalized
During devirtualization there is a chance that we will encounter virtual calls
inside a libfunc with IR. Most probably that virtual call will be another
libfunc. The problem is that since libfuncs with IR are external, there
is the possibility that we don't have all the possible targets. This
change set makes by default to include the virtual call as part of the
multiversioning when a libfunc is in the list of target functions or the caller
is a libfunc with IR.

Change-Id: I46c0a3ea96c67dc883bfb6788f7a9e19961bcb84
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/275643
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>