Resolve failures on CodeGenCXX/attr-likelihood-if-vs-builtin-expect.cpp
The test is supposed to check that [[likely]]/[[unlikely]] attribute results in
the same IR as __builtin_expect, but uses -O1 pipeline for that and we have
plenty of customizations, so the exact result in different.

I've modified the CHECKs to capture what xmain produces. The test is still doing
its job because we do verify that attributes behave in the same way as
__builtin_expect.

Change-Id: I224facf35c888dda3f0cae4d8ca874307a66e65e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/275771
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>