Teach HIRLocalityAnalysis to use loopcount_average/maximum metadata
HIRLocalityAnalysis is currently influenced by MAX_TC_EST when reasoning
about trip counts, but only in certain circumstances. For example, if
the trip count is not a constant, then MAX_TC_EST is clamped to
SymbolicConstTC, which is currently 100.

In other words, MAX_TC_EST on symbolic trip counts is interpreted with
somewhat low confidence.

In the case of OpenMP-parallelized loops the per-thread trip count is
typically symbolic even when the total trip count was a compile-time
constant because the number of threads is typically determined at
runtime. A consequence is that locality analysis almost always clamps
OpenMP loop trip counts to 100.

VPOParopt now computes conservatively correct loopcount_maximum and
loopcount_average metadata matching OpenMP scheduling. We can use these
as stronger hints, not clamped to 100, which helps locality analysis to
understand when OpenMP loops are likely to have high trip counts.

Change-Id: Ibb479bbebbab4bdef17e07ef6219d439b5a9239a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/276829
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>