[NFC] [DTrans - OpaquePtr] Extend debug trace messages for call parameter safety checks to consistently report argument number
Extends the debug dump method to take an optional callback method to allow additional
context information to be reported after the initial trace message reports the safety
message and value. This is to avoid the situation that occurred in the original DTrans
implementation where sometimes the information would be reported before the safety
message and sometimes after.

Change-Id: I6e141b6413d8a6b0e3480fb87067c23549c315be
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/276986
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>