CMPLRLLVM-23571: Fix for constant propagation in GEPs for multidimensional arrays
This change set enables the constant SCCP for multidimensional arrays.
The condition that checks if the data types matches for the parameters
was very conservative and ignored the case when a GEP is accessing a
multidimensional array. It produced a type mismatch and disabled the
constant propagation. This change set checks if there is a
mutidimensional array accessing the GEP, then keep iterating to make
sure it is accessing the right type.

Change-Id: I20d8543a91c9e3b87231c86db10fd3635e340ed0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/277598
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>