[SYCL] Fix OMM error after setting large WG sizes
This patch fixes out of memory error caused by large number
of global work size. It reproduces only on OpenCL FPGA Emu on the
specific machine. Now, if work group size wasn't specified in the
kernel source or IL, WGSize sets to {1, 1, 1}. It is the safest way to
get rid of the problem.

Change-Id: Ib70180f6018747943ce6ca27daacf7f0eb366602
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/277479
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: Vodopyanov, Dmitry <dmitry.vodopyanov@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>