CMPLRLLVM-23855: Teach HIR to accept but drop parallel_access metadata
HIRRegionIdentification skips loops with metadata not known to be
supported. VPOParopt recently started emitting
"llvm.loop.parallel_accesses" metadata, which can be considered
trivially supported in that they are safe to ignore or drop.

With this change, HIRRegionIdentification understands that this metadata
is supported instead of skipping such loops. However, the metadata is
ignored by loopopt and will not be preserved by HIRCodeGen.

This also fixes CMPLRLLVM-23847, a performance regression which resulted
from the new metadata causing HIRRegionIdentification to skip loops.

This is also required to resolve CMPLRLLVM-23998.

Change-Id: I3db1d052a4dcb40c4655ee308809c88a075ea931
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/277877
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>