[DTrans - OpaquePtr] Extend PtrTypeAnalyzer GEP processing to handle runtime dependent offsets into arrays
This is the first of several changesets needed to extend the information
stored by the PtrTypeAnalyzer for GEP instructions that index into nested
types to properly set the safety bits when the address is stored to memory
or passed to a function. In this change, an array that has a runtime
dependent index is marked with a special ElementPointee type in the
ValueTypeInfo object.

Change-Id: Iac31e7d2aac17aecffea942b188cbdc7a7f74aa9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/278164
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>