CMPLRLLVM-22930 and CMPLRLLVM-23571: Fix for compfail in CPU2017/627.cam4
This change set fixes the compfail in CPU2017/627.cam4 (CMPLRLLVM-22930).
The issue is that there is a GEP passed as actual parameter, but the pointer
operand is a BitCast. This produces a mismatch in types when generating the
constants during SCCP.

These changes also include the fix for CMPLRLLVM-23571, which is a regression
for polyhedron/ac,protein. The regression was caused by fixing CMPLRLLVM-22930.
Both changes are related.

Change-Id: I416de802d6e707290d8ffae254b1f322d1fad332
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/278171
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Huang, Zao <zao.huang@intel.com>