Teach NontemporalStore transformation to use ivdep_loop
This transformation requires understanding that a store in a loop can be
delayed until after the loop. In the case of 519.lbm_r we must show that
several stores to the same object are not writing to the same memory for
the duration of the loop. Simplified a bit:

    do i = 1..100:
      dst[i] = ...
      dst[i+1000] = ...

With 619.lbm_s OpenMP's iteration scheduling has the effect that the
trip count is not constant, making it less obvious that the stores don't
conflict:

    bound = __kmpc_runtime_for_static_init(...)
    do i = 1..bound:
      dst[i] = ...
      dst[i+1000] = ...

This change improves the legality analysis to show no conflicts in this
case. It takes advantage of the "ivdep loop" metadata (attached to
non-monotonic OpenMP loops) which asserts that there are no loop-carried
dependences. When the metadata is present the legality analysis can
reason about the stores' footprints only within an iteration rather than
across all iterations. This simplifies disambiguation: we only need to
show that i + 1000 != i.

Before querying alias analysis to perform the disambiguation, the
legality analysis now performs this simple constant difference check with
SCEV. In general alias analysis can disambiguate between non-overlapping
accesses to the same array with precise access sizes, but the cases in
619 seem too complex.

Change-Id: Icaaa646202bf3f2e412ff8e75940ee151e24eade
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/296501
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>
(cherry picked from commit 9443d374a648f3803a72569e916cb25c60c0bee1)
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308632
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Ng, John <john.ng@intel.com>
Reviewed-by: yaozhu <yao.zhu@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>