[Paropt][NFC] Move shared items privatization code to a new function pass
Preliminary change in preparation for enabling shared privatization with
nested regions. That requires refreshing nested work region nodes after
performing transformation, but it looks like this functionality is not
available for WRegionNode now. Moving privatization code to a separate
pass solves the problem because will force WRegionInfo to be regenerated.

Change-Id: Id13e5ec1e5e19ec6e2631d38541c06252b361e89
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/294972
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>