[Paropt][Dispatch] Handle cases when dispatch region has code other then a call to a function with variant.
This should handle cases like:

* Case 1: Pointer offset computation on use_device_ptr operand in dispatch regin
  #pragma omp target variant dispatch use_device_ptr(x)
  foo(x + 2)

* Case 2: Other non-variant calls in variant region
  #pragma omp target variant dispatch
  printf("%d\n", foo());

* Case 3: Subscript expressions on Fortran dope vectors in the region
  integer::a(:,:)
  !$omp target variant dispatch use_device_ptr(a)
  foo(a(:, :))

Note that this also means that use_device_ptr operands are privatized in the full region of
dispatch constructs, and not just the arguments.
So, for cases like this:

  y = &a[0];
  #pragma omp target variant dispatch use_device_ptr(y)
  y = foo(w)

"y" will be privatized for the dispatch "region", so the assignment of foo(w) to it will
not be visible after the dispatch region, and it will retain the value it had before the region.

Change-Id: Ie50926afea0a1277a8edf238126db861bfec12a9
JIRA: CMPLRLLVM-23934
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/295097
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>