CMPLRLLVM-241876: Insert casting to handle array to pointer, or vice-versa, in SCCP
This patch fixes the compfail when the arguments in a call poinst to the same
constant but the types are different. The idea is pretty simple, if there
formal argumnent is a pointer but the actual argument is an array, or vice-versa,
then we insert a BitCast instruction from the constant to the value. The constant
propagation will take care of propagate the rest.

This change set fixes the comfail in cpu2017/628.pop2 and
the runfail in specACCEL/557 CPUOCL.

Change-Id: Ia45e6da5767fd0ecd095043d10295a151b3c1002
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/296068
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>