CMPLRLLVM-23835: Cherry-pick d68bed0fa94e from the community
This change set is a cherry-pick from the community to fix a compfail.
The change set is here: https://reviews.llvm.org/rGd68bed0fa94e

This is the commit message from the community:

[SCCP] Handle bitcast of vector constants.

Vectors where all elements have the same known constant range are treated as a
single constant range in the lattice. When bitcasting such vectors, there is a
mis-match between the width of the lattice value (single constant range) and
the original operands (vector). Go to overdefined in that case.

Fixes PR47991.

Change-Id: Iec9718ce53c840dcc4559c20914efc70e72dfa9e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/295874
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>