[CMPLRLLVM-20406][6/6] Extend HIRSumWindowReuse to support intervening conditionals
This is the sixth and final change in a series to implement
HIRSumWindowReuse: it adds support for intervening conditionals within
the outer loop and is intended to achieve the gain reported in
CMPLRLLVM-11821.

The HIRAosToSoa change in
https://git-amr-2.devtools.intel.com/gerrit/275790/ is also needed to
see the full gain from this change, but without this change causes a
slight regression. To avoid this performance drop, it will be checked in
just before this change is.

Change-Id: I2f2791139b33131c46b1f9b1142e986e1ccd3203
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/272117
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>