[BranchProbabilityInfo] Get rid of MaxSuccIdx. NFC
This refactoring allows to eliminate the MaxSuccIdx map
proposed in the commit a7b662d0.
The idea is to remove probabilities for a block BB for
all its successors one by one from first, second, ...
till N-th until they are defined in Probs. This works
because probabilities for the block are set at once for
all its successors from number 0 to N-1 and the rest
are removed if there were stale probs.
The protected method setEdgeProbability(), which set
probabilities for individual successor, is removed.
This makes it clear that the probabilities are set in
bulk by the public method with the same name.

Reviewed By: kazu, MaskRay

Differential Revision: https://reviews.llvm.org/D90837

Change-Id: Ie6cb5f12052e000bea03dc319e0d25839bb5e959
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/298201
Reviewed-by: Link Ren <link.ren@intel.com>
Tested-by: Link Ren <link.ren@intel.com>