[BranchProbabilityInfo] Remove block handles in eraseBlock()
BranchProbabilityInfo::eraseBlock() is a public method and
can be called without deleting the block itself.
This method is made remove the correspondent tracking handle
from BranchProbabilityInfo::Handles along with
the probabilities of the block. Handles.erase() call is moved
to eraseBlock().
In setEdgeProbability() we need to add the block handle only once.

Reviewed By: kazu

Differential Revision: https://reviews.llvm.org/D90838

Change-Id: I1710b30bcfd94b911912fc8789fecc78b8490ac4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/298202
Reviewed-by: Link Ren <link.ren@intel.com>
Tested-by: Link Ren <link.ren@intel.com>