[1/2][NFCI][VPlan] Uplift HIR call vectorization codegen to be VPCallInst-based.
Currently HIR vectorizer's codegen always widens a call instruction
based on underlying HLInst knowledge. This patch changes this
behavior by uplifting vector call codegen to be based on decisions
and information captured in VPCallInstruction. The updates loosely
mimic the design used in VPlan LLVM-IR vectorizer codegen. We also
deprecate support for handling calls in mixed CG as a step towards
eventual retirement of the mixed CG approach. This is the first
patch in the series for the task to implement serialization feature
in HIR vectorizer.

Change-Id: I01dcd6fcd7e7f35f44abca9ffd37955fab2dfa5b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/298872
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>