[CMPLRLLVM-24433] [DTrans] Fix for DTransAnalysis not setting up MemcallInfo correctly for cases of element-pointee/non-element-pointee parameters
The delete field extension for deleting fields from nested types allowed for
memcpy calls using a structure field member and a pointer to an aggregate
type that was not a member of the structure. For example:
  %struct.test01a = type { i32, i32, i32, i32, i32 }
  %struct.test01b = type { i32, %struct.test01a }
A pointer to a %struct.test01a obect, and the field address of the
%struct.test01a field of %struct.test01b was allowed.

In this case, the MemCallInfo was set incorrectly to allow the delete
field transformation to adjust the size parameter if fields were deleted
from %struct.test01a.

Change-Id: I77c7a6db48fa83065d04661e9a0fc32c041e93cc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/296765
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>