CMPLRLLVM-24423: New pass manager changes order of functions in inline report
I have updated the LIT tests to match the order that the functions are
emitted in the classic inlining report with the new pass manager.
In the future, it might be useful to have a mode in which the functions
are emitted in alphabetical order, but this will suffice for now. One test
was updated to use a DAG check because it did not have INLINE statements
nested within COMPILE FUNC statements.

Change-Id: If6c7a9bfdd6d0775b4cb10db782000170eb28ca8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/296737
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>