[loopopt] HIRRuntimeDD: Do not ignore "ivdep loop" loops by default
Toward CMPLRLLVM-25806 and CMPLRLLVM-25991.

As a heuristic, loops marked with any kind of "ivdep" pragma are
considered unprofitable to multiversion. The intuition is that this
pragma tells us that there are no loop-carried dependencies, meaning
that multiversioning is unlikely to improve the DDG much.

However, VPOParopt recently began emitting "ivdep loop" markings to
preserve OpenMP "parallel for" semantics, and there are cases where
breaking loop-independent dependencies is important.

For example, in hwr_v2/Seismic~3d_smooth we need to disambiguate between
"in" and "out" allocatable arrays in order to perform unroll-and-jam.
Until the Fortran frontend can emit metadata to help disambiguate we
depend on multiversioning to disambiguate.

This change modifies the HIRRuntimeDD heuristics so that "ivdep"/"ivdep
back" still cause it to give up on the transformation, but "ivdep loop"
in particular is allowed.

Change-Id: I6a3779666e34cd51a5bdb4daf69de67d1126a877
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/300666
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>