[DTrans - OpaquePtr] DTransSafetyAnalyzer support for analyzing memcpy/memmove intrinsic calls
This adds analysis of memcpy/memmove calls to identify cases that are supported
by DTrans because the call has matching aggregate types for both pointer
parameters, and the size can be analyzed as covering a multiple of the aggregate
type size or a subset of fields. Unsupported cases will be marked with a
safety flag.

Change-Id: I8974ccbab67eafe828fb63a238677827bb153c22
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/297225
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>