CMPLRLLVM-24270: Ensure that stacksave/stackrestore brackets any inlined code if the caller has an OpenMp directive.
A previous Intel customization inhibited the hoisting of static allocas when
the caller had an OpenMp directive, but did not include the stacksave/stack-
restore brackets around the inlined code. Consequently, a stack overflow
occurred when the OpenMp directive was on a loop, because the static allocas
were executed on each iteration of the loop without being deallocated.

Change-Id: I7948a694933f6b6ddc9e7ef681bcd4b86da9496a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/297963
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>