[VPlan] Emit VPVectorTripCount instead of original upper bound in HIR path.
This is the first part of changes to eliminate new loop IV creation
when a normalized one exist in the vectorized loop.
1) Added two utility functions to detect normalized IV in the loop.
2) For HIR path where we always have normalized IV, implemented replacement
of the original trip count by vector trip count. Although that vector trip
count is not used for actual code generation at the moment, the change
will align VPlans for HIR and LLVM IR regarding loop iv.

Change-Id: I9da1fe34830d66a233eecab8a0b533b38d4eab6a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/299481
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>