[VPlan][HIR] Support instruction serialization in HIR vectorizer codegen.
This patch introduces a new serialization feature for VPlan HIR
vector codegen in order to handle instructions that cannot be
explicitly widened. This is achieved by generating a scalar
HLInst for each vector lane, after extracting appropriate operands
and collecting scalar results in the final vector. Predication is
handled on-the-fly by means of creating a HLIf using corresponding
lane's mask, and inserting serialized instruction to then branch.
This is the last patch in the series for the task to implement
serialization feature in HIR vectorizer.

Change-Id: I8b2ea3b8e8177e30c829a07ab912301365070630
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/298873
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>