[DTrans - OpaquePtr] PtrTypeAnalyzer support for partial pointer use idiom
Update the PtrTypeAnalyzer to ignore pointer types that are aliases when these
types are used for the partial pointer use idiom. This idiom is used in the
spec_qsort pointer copying where the value is copied in chunks. The pattern
matching code is ported from the implementation in LocalPointerAnalyzer to
the PtrTypeAnalyzer. Various changes made to the PtrTypeAnalyzer to use the
information.

Change-Id: I242cc9e6be344deedc68d0faca20792777a1f579
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/298743
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>