Add x86_amx type for intel AMX.
The x86_amx is used for AMX intrisics. <256 x i32> is bitcast to x86_amx when
it is used by AMX intrinsics, and x86_amx is bitcast to <256 x i32> when it
is used by load/store instruction. So amx intrinsics only operate on type x86_amx.
It can help to separate amx intrinsics from llvm IR instructions (+-*/).
Thank Craig for the idea. This patch depend on https://reviews.llvm.org/D87981.

Differential Revision: https://reviews.llvm.org/D91927

Change-Id: I905725a494227077a7d4c09c518d43f2cbca6ca1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/309220
Tested-by: Luo, Yuanke <yuanke.luo@intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>