[SYCL][ESIMD] - Disabled optimizations in esimd FE tests
CMPLRLLVM-24735.

Loop optimization pipeline was disabled for SYCL device compilation.
We have a '#pragma unroll' in one of the simd construsctors.
Because there is no loop unrolling pass in the pipeline, the
-transform-warning pass raises warning for the missed transformation.

Since tests that are changed in this patch are purely FE, I disabled
the whole optimization pipeline in those tests.

Change-Id: Icf419f402083af7e86363d022726296a50de9c40
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/299734
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Bobrovsky, Konstantin S <konstantin.s.bobrovsky@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>