Take into account an alignment from a mem intrinsic when expanding a struct copy
An alignment is specified for a source and a destination of a memory
transfer intrinsic in the InstCombinerImpl::SimplifyAnyMemTransfer
function. Take into account the specified alignment in the customization
which expands memcpy of a struct into a series of structure member
assignments. In common cases this alignment will be equal to the largest
default alignment among all structure member types. But in the case
when packed structure is processed we must take into account alignments
specified for src and dst pointers in memcpy because they can be less
than default alignment of the structure member types.

Change-Id: Ic5f178c663d28bad8641ab27761b8ce63bb308e6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/299910
Tested-by: Link Ren <link.ren@intel.com>
Reviewed-by: Link Ren <link.ren@intel.com>
(cherry picked from commit 7387946b6eb83309d356818f3356404d02981d8b)
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/301222