CMPLRLLVM-24772: Fix to not treat copy with AddRec header phi operand as IV update.
There is some logic in SSA deconstruction to keep IV update copies at the end
of the loop w.r.t other deconstructed phi copies to correctly deconstruct SSA
when IV is parsed a a blob. The logic was incorrectly recognizing a copy with
an AddRec header phi operand as an IV update resulting in incorrect order
of deconstructed copies.

Change-Id: Ide0571842c541820073aeb043e7b14b8ae5ddcf3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/300107
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: chenyang liu <chenyang.liu@intel.com>
Reviewed-by: Yu, Lina <lina.yu@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>