Added tests to verify that devicelib APIs have no 'noinline' attribute. (#2847)
Device compilers may (and should) honor 'noinline' attribute
even when optimizations are turned on. At the same time, we do not want
to prevent inlining of devicelib wrappers and fallback implementations,
since this may result in degraded performance. These tests verify that
the devicelib APIs are not marked with 'noinline' attribute.

Change-Id: I8a86b044975f70a7bee7e6deefdbe71cac1b4a43
Signed-off-by: Vyacheslav Zakharin <vyacheslav.p.zakharin@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/302354
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Jin, Ge <ge.jin@intel.com>
Tested-by: Jin, Ge <ge.jin@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>