Run SYCL tests with host device enabled
This commit turns on SYCL_ENABLE_HOST_DEVICE environment variable for
all SYCL tests and removes workarounds for the lack of host device
from several unit tests. This is done to avoid the need for further
xmain-specific modifications for future tests relying on the
spec-conformant behaviour of host device availability.

Change-Id: I2b534e2c9d373db0ede201608ebdd0fdad174cc0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/302053
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>
Tested-by: Semenov, Sergey <sergey.semenov@intel.com>
Reviewed-by: Pengfei Wang <pengfei.wang@intel.com>