[VPlan] Change the loop IV comparison from EQ/NE to UGE/ULT.
The change is possible due to we have normalized IV (in scalar loop,
it starts from 0 and increments by 1) so the end iteration is perfectly
detected by ULT/UGE.
The change is made to have a loop latch comparison in the vector loop
really uniform: the vector loop IV starts from 0 and increments by VF.
So having e.g. ULT and comparing with vector trip count (which is mutiply
of VF) on each vector iteration we will have the same results of comparison
for all vector lanes.
The new check is more suitable for masked mode loop creation. Because the
bottom condition of the masked mode loop will look like the following:
   %be_cond = icmp ult %iv, %loop.orig.tc

Essentially, the original trip count is used here. There is no need to
calculate the vector trip count for this loop. It's a run-time calculation,
thus it's an additional overhead. Then, the %iv is incremented by VF and,
in general, it may not ever become equal to the original tc - it can be
either less or greater of the original tc.

Change-Id: I5a9f064ddf80dd4a1952d8c90c8da0dd903e68e5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/302229
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>