[NFCI][VPlan] Correction in DA to process latch condition which compares with VectorTripCount.
If any of operands is VectorTripCountCalculation and condition is latch
condition we consider it as uniform. Because even the induction which
is compared with VectorTripCount is not uniform all its lanes contain
values in {x+0:x+VF-1} range, and x is started from 0 and incremented by
VF (we always have normalized induction). So the condition will be
true/false for all lanes simultaneously.

Change-Id: I44670d3e25a1480e4d16f9ac9fd4eb32abdf1884
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/302375
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>