[DTrans][SOAToAOS] Remove dtrans-soatoaos-ignore-classinfo option that is used for lit tests
dtrans-soatoaos-ignore-classinfo option was added to only be used in LIT tests,
which verify that soatoaos transformation is triggered, to ignore class-info.
When this option is true, soatoaos transformation ignores the expected
functionality of array member functions.
This option is removed completely and LIT tests are fixed to have
valid class-info with member functions that have expected functionality.

Change-Id: I6e58a0264ec2262f40af091fd18ef320b7249da7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/303215
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>