[HIR][VPlan] Tag vectorized HLLoops as auto/simd-vectorized in HIR dumps.
Currently there is no direct means to identify vectorization effects
on a HLLoop apart from checking the loop's step or explicit use of
VectorType in instructions inside the loop. This patch adds a new
tagging mechanism in HLLoop to indicate if the loop was auto/simd-
vectorized by VPlan. This tag is accordingly reflected whenever the
loop is printed.

Change-Id: I4810d506a9d67feefe071b1db655031d1396d504
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/303233
Reviewed-by: Tian, Xinmin <xinmin.tian@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>