[loopopt] Adjust for "ivdep loop" in DDTests
Previously "#pragma ivdep loop" and "#pragma ivdep back" were treated
identically, setting direction vectors to "=" except where the distance
was known to be small. (Less than 128.) This initial behavior was
inspired by icc, and is arguably a little vectorizer-centric.

With this change, "ivdep loop" is interpreted more aggressively. The
"ivdep back" behavior is not altered by this change. In the case of
"ivdep loop," the direction is masked with "=". This adjustment is a
true conjunction rather than a replacement.  For example, if before an
"ivdep loop" adjustment the direction was known to be "<", afterward it
will be "NONE" rather than "=".

To avoid creating direction vectors with an empty/NONE outer direction,
only innermost loops are candidates for "ivdep loop" adjustment. Such
direction vectors would not necessarily be interpreted correctly by
other parts of DD.

Change-Id: Ib8feec1d8ac39c25a16ca8fe62623cd233502352
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/303560
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Sergey Kanaev <sergey.kanaev@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/308633
Reviewed-by: Ng, John <john.ng@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>