CMPLRLLVM-11587: Eliminate identical nested ZTT during loop formation phase.
In some cases we can have identical nested ifs in the incoming IR. This
is happening for 503.bwaves. There are multiple ScalarOpt passes like
InstCombine and SimplifyCFG which can eliminate redundant conditions
but their functionality is limited and dependent on how complicated the
CFG is when these passes are triggered in the pipeline.

As a last resort, I am special casing this in the LoopOpt framework.

Change-Id: I52380d4bfad407d0e548e09e46a1c64c4218c780
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/303436
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>