[VPlan][HIR] Allow mixed CG for loops containing ignored calls.
We recently disabled call vectorization in mixed CG approach (via
a bailout) as part of efforts to handle them only in VPValue-based
CG approach. This bailout resulted in performance regressions in a
key benchmark where the hot loop required mixed CG and contained
calls to lifetime intrinsics that are ignored during codegen. This
patch adds a temporary mechanism to relax the bailout for calls
that will be ignored, thereby allowing trivial mixed CG for such
cases. The mechanism is expected to be removed when VPValue-based
CG is updated to handle liveouts or when mixed CG is retired.

JIRA : CMPLRLLVM-25257

Change-Id: Idd088c540591d86ceb61fe298324c792534cb0fd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/303838
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>