[NFC] Update openmp tests which use archives
There are a few OpenMP offloading tests that use archives.  The behavior
for handling archives has changed for offloading, and the tests need
to be updated to match.

Update for xmain-cand, to use aoo instead of a for unbundling

Change-Id: I38d094cbca64135e659d0f5397e2a9622c04403e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/304035
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
(cherry picked from commit 65f21a62cec6dff4b7dc010aaf3934ea9ba661be)
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/304055
Tested-by: Link Ren <link.ren@intel.com>
Reviewed-by: Link Ren <link.ren@intel.com>