[SYCL] [FPGA] Disable swapping dimension values of WorkGroupAttr arguments for semantic checks (#2962)
This is a followup comments on https://github.com/intel/llvm/pull/2906#discussion_r546940763
and https://github.com/intel/llvm/pull/2906#discussion_r548524994.

Currently
For a SYCLDevice WorkGroupAttr arguments are reversed.
  if (S.getLangOpts().SYCLIsDevice)
    std::swap(XDimExpr, ZDimExpr);

There are few problems with this approach (swapping order in AST):
1. A user who writes the order as X, Y, Z in their source will get Z, Y, X when
   they pretty print the AST.
2. The same will happen when dumping the AST to text or JSON or
   generating diagnostic messages since the fist and third arguments of
   WorkGroupAttr are reversed.
3. This is very confusing to see the different order than the oroginal
   source codes.

max_work_group_size() follows same rules as reqd_work_group_size() in Sema and LLVM IR generation.

This patch keeps the arguments in the same order as they were parsed in doing semantic checks,
and when performing code generation, we would change the argument order when lowering to
LLVM IR and not when creating the semantic attribute.

Change-Id: I41a0d5f87b50faa2b283f645e75e6aa615070583
Signed-off-by: Soumi Manna <soumi.manna@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/304966
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>