[Driver][SYCL][FPGA] Allow for multiple AOCX device binaries for FPGA (#2909)
Initial implementation only allowed for a single AOCR/AOCX binary to
be used in a single FPGA compilation.  This restriction is being lifted
and we are allowing multiple FPGA device images to be added to the
final executable.

Allow for list file creation for aocx/aocr files with update extensions
when unbundling the archives.  To handle this, we are also enabling the
rename functionality for file-table-tform

Improve 'aoc' check in FPGA test for systems that have aoc installed
remove fpga_device_link_diag.cpp test, as we no longer do the spir check

Add TODO for improvement of output type and naming for aocx/aocr unbundles

This is a cherry-pick of: https://github.com/intel/llvm/commit/6ea38f0f1f7ae533f6a773d40e6d57f6ba9a08da

Change-Id: I77012defd394fcb011299da2816b27c9ff4bc973
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/305506
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>